requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
psycopg2-binary
fastapi
uvicorn
streamlit
pydantic
python-multipart
requests
plotly